+++ BingolSinema/Models/Bilet.cs
@@ -9,7 +9,7 @@ public class Bilet
     [Key]
     public int BiletID { get; set; }
     public required int RezervasyonID { get; set; }
-    public required double Fiyat { get; set; }
+    public required int Fiyat { get; set; }
 
     // Diğer özellikler eklenebilir
 }
